+++ test/test036.js
@@ -3,10 +3,18 @@ if(typeof exports === 'object') {
 	var alasql = require('..');
 };
 
-describe('Test 36', function() {
-	it('INSERT INTO FROM', function(done){
 
-		var db = new alasql.Database("db");
+describe('Test 36', function() {
+	var db;
+
+	it('1. Create Database', function(done){
+		alasql.options.modifier = undefined;
+		db = new alasql.Database("db");
+		done();
+	});
+
+	it('2. INSERT INTO FROM 1', function(done){
+
 		db.exec('CREATE TABLE test1 (a STRING)');
 
 		var sql = "INSERT INTO test1 (a) VALUES ('Alpha'), ('Beta'), ('Gamma'), ('Delta'), ('Epsilon')";
@@ -14,6 +22,10 @@ describe('Test 36', function() {
 		
 		var sql = "SELECT VALUE COUNT(*) FROM test1";
 		assert.equal(5,db.exec(sql));
+
+		done();
+	});
+	it('3. INSERT INTO FROM 2', function(done){
 
 		db.exec('CREATE TABLE test2 (a STRING)');
 
@@ -23,6 +35,9 @@ describe('Test 36', function() {
 		var res = db.exec("SELECT * FROM test2");
 		assert.deepEqual([{a:'Gamma'}], res);
 
+		done();
+	});
+	it('4. INSERT INTO FROM 3', function(done){
 		db.exec('CREATE TABLE test3 (a STRING)');
 
 		var sql = "INSERT INTO test3 SELECT * FROM test1 WHERE a NOT LIKE '%e%'";
@@ -33,4 +48,8 @@ describe('Test 36', function() {
 
 		done();
 	});
+	it('99. Drop database', function(done){
+		done();
+	});
+
 });